Create Agar_io.py
Python will have trouble running a file with a `.io` file extension when it expects Python scripts to have the file extension `.py`.

You should make Particles and Bots be subclasses of `Rect` and get rid of .cells and .attacker.  You can then do `if player.intersects(particle)` or `if player.intersects(bot)`.

The globals make this code impossible to understand and debug.  You really need to get rid of them.  If Intro has links to the Player, all particles, and all bots then NO globals are needed.